requirements.txt
pandas
numpy
sqlalchemy
yfinance
sec-edgar-downloader
pyyaml
networkx
requests
lxml